fix: alter project root name
