+++ backend/CurrencyConvertor/settings.py
@@ -22,7 +22,6 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'api',
-    'currencys',
 ]
 
 MIDDLEWARE = [
